Fixed error in random_ysos() where ysos were being generated outside of the coverage map. This was due to pixels being centered at the coordinates from wcs_object.all_pix2world() and the random perturbation around the centre would push some points into a neighbouring pixel.
Added new mode to random_ysos(): sphere_binomial.
This method randomly generates random coordinates on the surface
of the celestial sphere that are within the coverage map's extreme
bounds. YSOs are iteratively added until the desired number of YSOs
have been added to the map within the covered part of the map.